requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
blinker==1.4
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
coloredlogs==15.0.1
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
fastapi==0.87.0
filetype==1.2.0
flatbuffers==23.1.21
frozenlist==1.3.3
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h11==0.14.0
h5py==3.8.0
httplib2==0.20.2
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.25.0
importlib-metadata==4.6.4
jeepney==0.7.1
keras==2.11.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==15.0.6.1
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.2
more-itertools==8.10.0
mpmath==1.2.1
multidict==6.0.4
netifaces==0.11.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.0
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==23.0
Pillow==9.3.0
pooch==1.6.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.4
PyGObject==3.42.1
PyJWT==2.3.0
PyMatting==1.1.8
pyparsing==2.4.7
python-apt==2.3.0+ubuntu2.1
python-multipart==0.0.5
PyWavelets==1.4.1
PyYAML==5.4.1
rembg==2.0.30
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.3
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
sympy==1.11.1
systemd-python==234
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
tifffile==2023.1.23.1
tqdm==4.64.1
typing_extensions==4.4.0
ubuntu-advantage-tools==27.12
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.14
uvicorn==0.20.0
wadllib==1.3.6
watchdog==2.1.9
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.2
zipp==1.0.0
